requirements.txt
Django==5.1.1
whitenoise==6.7.0
gunicorn==22.0.0
django-distill==3.2.4
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
django-filter==24.2
